requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic-settings
jinja2
python-multipart
requests
